src/be/requirement_extration/requirement_extration/requirements.txt
azure-storage-blob==12.23.1
azure-search-documents==11.5.2
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.26.4
tqdm==4.66.6
ipywidgets==8.0.4
python-dotenv~=1.0.1
pymilvus==2.4.9
transformers==4.46.1
torch==2.5.1
spacy==3.7.5
pdfminer.six==20240706
PyMuPDF==1.24.13
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
openpyxl==3.1.5
XlsxWriter==3.2.0
markdown2==2.5.1
openai==1.55.0
python-Levenshtein==0.26.1
fuzzywuzzy==0.18.0
sentence-transformers==3.3.1
tiktoken==0.8.0
aiofiles==24.1.0
langchain==0.3.25
langchain-cli==0.0.36
langchain-community==0.3.23
langchain-core==0.3.60
